+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # What's New!
 
+**Changelog** (07-02-2025)
+
+- OPHUB Add Kernel 6.x.x
+- Add Source Packages By openwrt.ai
+- Use Deafult Font Terminal
+- Add AdGuard Home
+
 **Changelog** (05-02-2025)
 
 - Add Theme Material3